Module fuzzy_logic::functions

This module defines types and structures for fuzzy logic functions.

Module contains implementation of membership functions and defuzzification functions. Also contains factory methods to create most used functions.

Structs

DefuzzFactory
Defines methods to create most used defuzzification functions.

MembershipFactory
Defines methods to create most used membership functions.

Type Definitions

DefuzzFunc
Used to defuzzificate the fuzzy logic inference result. All defuzzification functions must be this type.

MembershipFunction
Used to calculate the membership of the given item. All membership functions must be this type.