Module fuzzy_logic::ops

This module contains implementation of the fuzzy logic operations.

Fuzzy set operations and fuzzy logic operations are defined here.

User can implement his own operations by implementing LogicOps or SetOps traits.

Structs

MinMaxOps
Implementation of commonly used minimax set operations.

ZadehOps
Implementation of commonly used Zadeh fuzzy logic operations.

Traits

LogicOps
Abstraction over fuzzy logic operations. Doesn't contain default implementation.

SetOps
Abstraction over set operations. Doesn't contain default implementation.